Add mmap syscall
Add mmap() range check

Change MMAP_ANON to -1 to make it an exclusive flag.

Add check on mmap syscall so that it checks if a range of pages are
allocated instead of only a single page.

Move stack start address to a higher location in order to make space for
a potential heap.